Add validators for History / Snapshot Retention Policy (#259)
## Summary

<!--- HINT: Replace #nnn with corresponding Issue number, if you are
fixing an existing issue -->

Supports defining a policy to control the versions of snapshots tables
to retain.
The policy will support 2 types of configurations:
1. Time-based (e.g. 3 days)
2. Count-based (e.g. 10 versions)

It can also support a combination of the 2 policies, in which case it
will retain versions that fall under both categories (e.g. only keep
versions newer than 3 days AND within 10 versions).

This PR also refactors the PolicySpecValidator to accurately depict that
it's referencing to partition-based retention of tables. In the future,
we should refactor these policy classes to follow some `Policy`
interface for better standardization for future policy support
(duplicated interfaces between Replication, Partition Retention, Retain
Snapshot).

This PR also defines the maximums for snapshot retention defaults (TODO:
make this configurable/static?)

**The current maximums defined is 3 days and 100 versions**


## Changes
- [x] Client-facing API Changes
- [x] Internal API Changes
- [ ] Bug Fixes
- [x] New Features
- [ ] Performance Improvements
- [ ] Code Style
- [ ] Refactoring
- [ ] Documentation
- [x] Tests

For all the boxes checked, please include additional details of the
changes made in this pull request.

## Testing Done
<!--- Check any relevant boxes with "x" -->

- [x] Manually Tested on local docker setup. Please include commands
ran, and their output.
- [x] Added new tests for the changes made.
- [x] Updated existing tests to reflect the changes made.
- [ ] No tests added or updated. Please explain why. If unsure, please
feel free to ask for help.
- [ ] Some other form of testing like staging or soak time in
production. Please explain.

Ran on local docker and tested using postman by creating a new table
with the following:
```
{
    "tableId": "t2",
    "databaseId": "d3",
    "baseTableVersion": "INITIAL_VERSION",
    "clusterId": "LocalFSCluster",
    "schema": "{\"type\": \"struct\", \"fields\": [{\"id\": 1,\"required\": true,\"name\": \"id\",\"type\": \"string\"},{\"id\": 2,\"required\": true,\"name\": \"name\",\"type\": \"string\"},{\"id\": 3,\"required\": true,\"name\": \"ts\",\"type\": \"timestamp\"}]}",
    "tableProperties": {
        "key": "value"
    },
    "policies": {
        "sharingEnabled": "true",
        "history": {"maxAge": 1, "granularity": "DAY", "versions": 2}
    }
}
```
Tested getTable after and confirmed the policy was showing correctly.

Also tested the negative cases (high version count, negative numbers)
which worked as well.

For all the boxes checked, include a detailed description of the testing
done for the changes made in this pull request.

# Additional Information

- [ ] Breaking Changes
- [ ] Deprecations
- [ ] Large PR broken into smaller PRs, and PR plan linked in the
description.

For all the boxes checked, include additional details of the changes
made in this pull request.