Bumps OPA to version 1.0 release to fix integration tests and docker setup (#275)
## Summary

<!--- HINT: Replace #nnn with corresponding Issue number, if you are
fixing an existing issue -->

[Issue](https://github.com/linkedin/openhouse/issues/#nnn) Briefly
discuss the summary of the changes made in this pull request in 2-3
lines.

OPA recently released version 1.0.0 which is not backwards compatible.
Since our docker images were using the latest versions of OPA, there
were some changes needed in order to be compatible with the released
major version.
https://github.com/open-policy-agent/opa/releases/tag/v1.0.0

Opted to fix-forward rather than revert to 0.70.0 as it seems 1.0.0 is a
significant release which should be supported for longer term.

## Changes

- [ ] Client-facing API Changes
- [ ] Internal API Changes
- [x] Bug Fixes
- [ ] New Features
- [ ] Performance Improvements
- [ ] Code Style
- [ ] Refactoring
- [ ] Documentation
- [x] Tests

Ensures that the GHA CI/CD will wait for build-and-test to pass before
releasing. Although PRs are validated, there was still a miss when there
are docker/GHA version releases in between the last PR CI validation run
and the merge time

Validated the rules work using the OPA 1.0 binary and followed the
compatibility guide
https://www.openpolicyagent.org/docs/latest/v0-compatibility/

Needed to explicitly bind the port of the server to `0.0.0.0` as it now
defaults to localhost, causing an issue when being accessed by other
containers.

## Testing Done
<!--- Check any relevant boxes with "x" -->

- [x] Manually Tested on local docker setup. Please include commands
ran, and their output.
- [ ] Added new tests for the changes made.
- [ ] Updated existing tests to reflect the changes made.
- [ ] No tests added or updated. Please explain why. If unsure, please
feel free to ask for help.
- [x] Some other form of testing like staging or soak time in
production. Please explain.

Tested that this works e2e by running the integration tests and running
the local docker setup.


# Additional Information

- [ ] Breaking Changes
- [ ] Deprecations
- [ ] Large PR broken into smaller PRs, and PR plan linked in the
description.

For all the boxes checked, include additional details of the changes
made in this pull request.

# Additional Information

- [ ] Breaking Changes
- [ ] Deprecations
- [ ] Large PR broken into smaller PRs, and PR plan linked in the
description.

For all the boxes checked, include additional details of the changes
made in this pull request.