Always enforce a ttl on snapshot expiration jobs even if unconfigured (#281)
## Summary

<!--- HINT: Replace #nnn with corresponding Issue number, if you are
fixing an existing issue -->

[Issue](https://github.com/linkedin/openhouse/issues/#nnn)] Briefly
discuss the summary of the changes made in this
pull request in 2-3 lines.


Currently it is possible to define tables with an version count, and the
snapshot expiration job will respect the version count instead of
enforcing a TTL. With this change the snapshot expiration job will
**always** enforce a TTL, defaulted currently to 3 days. This cannot be
overridden by users unless they explicitly define a TTL that is greater
than 3 days (but is currently the maximum is 3 days due to the history
configuration validator)
https://github.com/linkedin/openhouse/blob/7f66a730ec7d611335917e29bfed01050090c397/services/tables/src/main/java/com/linkedin/openhouse/tables/api/validator/impl/HistoryPolicySpecValidator.java#L57

Also fix some bugs where `TableSnapshotExpirationTask` is not sending
properties correctly from the JobScheduler

## Changes

- [ ] Client-facing API Changes
- [ ] Internal API Changes
- [ ] Bug Fixes
- [ ] New Features
- [ ] Performance Improvements
- [ ] Code Style
- [ ] Refactoring
- [ ] Documentation
- [ ] Tests

For all the boxes checked, please include additional details of the
changes made in this pull request.

## Testing Done
<!--- Check any relevant boxes with "x" -->

- [ ] Manually Tested on local docker setup. Please include commands
ran, and their output.
- [ ] Added new tests for the changes made.
- [x] Updated existing tests to reflect the changes made.
- [ ] No tests added or updated. Please explain why. If unsure, please
feel free to ask for help.
- [ ] Some other form of testing like staging or soak time in
production. Please explain.

Tested E2E with acceptance tests from local

For all the boxes checked, include a detailed description of the testing
done for the changes made in this pull request.

# Additional Information

- [ ] Breaking Changes
- [ ] Deprecations
- [ ] Large PR broken into smaller PRs, and PR plan linked in the
description.

For all the boxes checked, include additional details of the changes
made in this pull request.