PR1 (nullability bug): test structure prior to feature change (#290)
## Summary
problem: 
1. OpenHouseITest does not expose public methods to configure the spark
catalog to non-default settings. in
[PR2](https://github.com/linkedin/openhouse/pull/288/files), i must add
non-default settings when using OpenHouseITest
2. there isn't a CTAS test already for spark with catalog integration
(although there is one for spark and catalog but without being
integrated). in
[PR2](https://github.com/linkedin/openhouse/pull/288/files), I will
update the existing CTAS to isolate and validate a before/after of my
change.

solution: we refactor openhouseitest to allow for a builder to be
returned and configurable, and then add a new CTAS test for existing
behavior

PR2 after PR1 is merged will look like this:
https://github.com/cbb330/openhouse/pull/5/files

## Changes

- [ ] Client-facing API Changes
- [ ] Internal API Changes
- [ ] Bug Fixes
- [ ] New Features
- [ ] Performance Improvements
- [ ] Code Style
- [X] Refactoring
- [ ] Documentation
- [X] Tests

## Testing Done
<!--- Check any relevant boxes with "x" -->

- [ ] Manually Tested on local docker setup. Please include commands
ran, and their output.
- [X] Added new tests for the changes made.
- [X] Updated existing tests to reflect the changes made.
- [ ] No tests added or updated. Please explain why. If unsure, please
feel free to ask for help.
- [ ] Some other form of testing like staging or soak time in
production. Please explain.

# Additional Information

- [ ] Breaking Changes
- [ ] Deprecations
- [ ] Large PR broken into smaller PRs, and PR plan linked in the
description.