Support snapshots configurations in stats collector, fix bugs where snapshot retention didn't support past DAY format, fix false positive edge case in snapshot event (#291)
## Summary

<!--- HINT: Replace #nnn with corresponding Issue number, if you are
fixing an existing issue -->

[Issue](https://github.com/linkedin/openhouse/issues/#nnn) Briefly
discuss the summary of the changes made in this
pull request in 2-3 lines.

We want to be able to emit events with user tables that have the history
configuration for them so that monitoring systems can accurately detect
when snapshot expiration fails for configured tables.
This PR also does some refactoring around parsing policy JSONs to be
more robust and easier to manage as more policies are added.

This PR also fixes a bug where not all date granularities were supported
in retention, since `TimeUnit` is maxed granularity at DAYS when
technically the snapshot policy can be kept at higher granularities of
MONTH and YEAR, although we currently still limit it to 3 DAYS.

We also add the `secondOldestSnapshot` field to the IcebergTableStats to
handle the edge case where a table has very infrequent writes, this can
cause snapshots to fall out of retention policy but not yet be cleaned
up, if the secondOldestSnapshot is still recent then we should avoid
being alerted.

## Changes

- [ ] Client-facing API Changes
- [x] Internal API Changes
- [ ] Bug Fixes
- [ ] New Features
- [ ] Performance Improvements
- [ ] Code Style
- [x] Refactoring
- [ ] Documentation
- [ ] Tests

For all the boxes checked, please include additional details of the
changes made in this pull request.

## Testing Done
<!--- Check any relevant boxes with "x" -->

- [ ] Manually Tested on local docker setup. Please include commands
ran, and their output.
- [ ] Added new tests for the changes made.
- [ ] Updated existing tests to reflect the changes made.
- [ ] No tests added or updated. Please explain why. If unsure, please
feel free to ask for help.
- [ ] Some other form of testing like staging or soak time in
production. Please explain.

For all the boxes checked, include a detailed description of the testing
done for the changes made in this pull request.

Unit tests, tested snapshot version and ran tablestatscollector job in a
cluster on Spark, did not see any errors.


# Additional Information

- [ ] Breaking Changes
- [ ] Deprecations
- [ ] Large PR broken into smaller PRs, and PR plan linked in the
description.

For all the boxes checked, include additional details of the changes
made in this pull request.