SQL support for unsetting replication policy (#297)
## Summary

Replication policy allows users to set configs for table replication.
While updates can be done on it, there is not provision to delete the
config in case users want to disable the replication process.

This PR introduces SQL support for purging the replication policy.
Syntax:
`ALTER table <db>.<table> UNSET policy (REPLICATION)`

The syntax is inline with iceberg support for unsetting the table
properties.
ref: https://iceberg.apache.org/docs/1.5.1/spark-ddl/#alter-table

## Changes

- [x] Client-facing API Changes
- [ ] Internal API Changes
- [ ] Bug Fixes
- [x] New Features
- [ ] Performance Improvements
- [ ] Code Style
- [ ] Refactoring
- [ ] Documentation
- [ ] Tests

For all the boxes checked, please include additional details of the
changes made in this pull request.

## Testing Done

Docker testing steps:
1. Create table with replication, retention policy using alter table
statement
2. Unset replication policy. Show tblProperties has policy Json as:
```
|policies                                  |{
  "retention": {
    "count": 20,
    "granularity": "DAY",
    "columnPattern": {
      "columnName": "name",
      "pattern": "'yyyy-MM-dd-HH-mm'"
    }
  },
  "sharingEnabled": false,
  "replication": {
    "config": []
  }
}|
```
3. Set table replication again. Check tblProperties: 
```
|policies                                  |{
  "retention": {
    "count": 20,
    "granularity": "DAY",
    "columnPattern": {
      "columnName": "name",
      "pattern": "'yyyy-MM-dd'"
    }
  },
  "sharingEnabled": false,
  "replication": {
    "config": [
      {
        "destination": "'WAR'",
        "interval": "12H",
        "cronSchedule": "0 30 13/12 ? * * *"
      }
    ]
  }
}|
``` 
4. Update retention policy to validate any other policy can be updated
as expected.

<!--- Check any relevant boxes with "x" -->

- [ ] Manually Tested on local docker setup. Please include commands
ran, and their output.
- [x] Added new tests for the changes made.
- [x] Updated existing tests to reflect the changes made.
- [ ] No tests added or updated. Please explain why. If unsure, please
feel free to ask for help.
- [ ] Some other form of testing like staging or soak time in
production. Please explain.

For all the boxes checked, include a detailed description of the testing
done for the changes made in this pull request.

# Additional Information

- [ ] Breaking Changes
- [ ] Deprecations
- [ ] Large PR broken into smaller PRs, and PR plan linked in the
description.

For all the boxes checked, include additional details of the changes
made in this pull request.