Allow extension of preserved key checker to add read-only keys during table creation (#303)
## Summary

<!--- HINT: Replace #nnn with corresponding Issue number, if you are
fixing an existing issue -->

[Issue](https://github.com/linkedin/openhouse/issues/#nnn) Briefly
discuss the summary of the changes made in this
pull request in 2-3 lines.

The `PreservedKeyChecker` will automatically filter any preserved keys
when creating an OH table (e.g. `BasePreservedKeyChecker` filters keys
prefixed with `openhouse.`), as the expectation is that preserved keys
are only written by the OH server.

However, there are some use-cases where it becomes beneficial for users
to extend preserved table properties to have read-only table properties
that can be defined by users during table creation. This PR does a small
change to the PreservedKeyChecker to allow for adding read-only keys
only when the table is created.

One usecase is when a user owns a platform of tables which may need an
identifier e.g. `openhouse.platform`, which should be set during
creation time for any custom handling but we do not want a modification
of this key.


## Changes

- [ ] Client-facing API Changes
- [x] Internal API Changes
- [ ] Bug Fixes
- [x] New Features
- [ ] Performance Improvements
- [ ] Code Style
- [x] Refactoring
- [ ] Documentation
- [ ] Tests

For all the boxes checked, please include additional details of the
changes made in this pull request.

## Testing Done
<!--- Check any relevant boxes with "x" -->

- [ ] Manually Tested on local docker setup. Please include commands
ran, and their output.
- [ ] Added new tests for the changes made.
- [ ] Updated existing tests to reflect the changes made.
- [ ] No tests added or updated. Please explain why. If unsure, please
feel free to ask for help.
- [ ] Some other form of testing like staging or soak time in
production. Please explain.

For all the boxes checked, include a detailed description of the testing
done for the changes made in this pull request.

Added unit tests

# Additional Information

- [ ] Breaking Changes
- [ ] Deprecations
- [ ] Large PR broken into smaller PRs, and PR plan linked in the
description.

For all the boxes checked, include additional details of the changes
made in this pull request.