Skip Preserved Property check with Primary->Primary Table Replication (#328)
## Summary

<!--- HINT: Replace #nnn with corresponding Issue number, if you are
fixing an existing issue -->

[Issue](https://github.com/linkedin/openhouse/issues/#nnn)] Briefly
discuss the summary of the changes made in this
pull request in 2-3 lines.

## Changes

- [ ] Client-facing API Changes
- [ ] Internal API Changes
- [ ] Bug Fixes
- [ ] New Features
- [ ] Performance Improvements
- [ ] Code Style
- [ ] Refactoring
- [ ] Documentation
- [ ] Tests

For all the boxes checked, please include additional details of the
changes made in this pull request.

Primary->Primary Table Replication will still provoke the
preservedTableProperty check. Since Primary->Primary replication can be
used for backfilling purposes, this will block the snapshot commit as
the preserved table properties of both tables will differ (e.g.
createdTime).

To handle this, we need to treat Primary->Primary table replication
similarly to Primary->Replica when one of the primary tables were
created as a replica table, but still maintain the old check for
backwards compatibility. Primary tables created for this purpose and
newly created replica tables will have the property
`OPENHOUSE_IS_TABLE_REPLICATED_KEY` set during table creation.


## Testing Done
<!--- Check any relevant boxes with "x" -->

- [ ] Manually Tested on local docker setup. Please include commands
ran, and their output.
- [ ] Added new tests for the changes made.
- [ ] Updated existing tests to reflect the changes made.
- [ ] No tests added or updated. Please explain why. If unsure, please
feel free to ask for help.
- [ ] Some other form of testing like staging or soak time in
production. Please explain.

For all the boxes checked, include a detailed description of the testing
done for the changes made in this pull request.

Unit tests

# Additional Information

- [ ] Breaking Changes
- [ ] Deprecations
- [ ] Large PR broken into smaller PRs, and PR plan linked in the
description.

For all the boxes checked, include additional details of the changes
made in this pull request.